+++ bundle/src/test/java/com/adobe/acs/commons/models/injectors/impl/HierarchicalPagePropertyInjectorTest.java
@@ -17,10 +17,11 @@
  */
 package com.adobe.acs.commons.models.injectors.impl;
 
-import com.adobe.acs.commons.models.injectors.annotation.HierarchicalPageProperty;
 import com.adobe.acs.commons.models.injectors.annotation.impl.HierarchicalPagePropertyAnnotationProcessorFactory;
 import com.adobe.acs.commons.models.injectors.impl.model.TestHierarchicalPagePropertiesModel;
+import com.adobe.acs.commons.models.injectors.impl.model.TestPagePropertiesModel;
 import com.adobe.acs.commons.models.injectors.impl.model.impl.TestHierarchicalPagePropertiesModelModelImpl;
+import com.adobe.acs.commons.models.injectors.impl.model.impl.TestPagePropertiesModelModelImpl;
 import io.wcm.testing.mock.aem.junit.AemContext;
 import org.apache.sling.api.resource.Resource;
 import org.apache.sling.models.spi.Injector;
@@ -32,8 +33,7 @@ import org.junit.runner.RunWith;
 import org.mockito.InjectMocks;
 import org.mockito.junit.MockitoJUnitRunner;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNotNull;
+import static org.junit.Assert.*;
 
 @RunWith(MockitoJUnitRunner.class)
 public class HierarchicalPagePropertyInjectorTest {
@@ -43,7 +43,8 @@ public class HierarchicalPagePropertyInjectorTest {
 
     @InjectMocks
     private HierarchicalPagePropertyInjector injector;
-    private TestHierarchicalPagePropertiesModel adapted;
+    private TestHierarchicalPagePropertiesModel hierarchicalModel;
+    private TestPagePropertiesModel pageModel;
 
     @Before
     public void setUp() throws Exception {
@@ -53,19 +54,24 @@ public class HierarchicalPagePropertyInjectorTest {
         context.registerService(Injector.class, injector);
         context.registerService(StaticInjectAnnotationProcessorFactory.class, new HierarchicalPagePropertyAnnotationProcessorFactory());
         context.addModelsForClasses(TestHierarchicalPagePropertiesModelModelImpl.class);
+        context.addModelsForClasses(TestPagePropertiesModelModelImpl.class);
 
         Resource adaptable = context.request().getResource();
-        adapted = adaptable.adaptTo(TestHierarchicalPagePropertiesModel.class);
+        hierarchicalModel = adaptable.adaptTo(TestHierarchicalPagePropertiesModel.class);
+        pageModel = adaptable.adaptTo(TestPagePropertiesModel.class);
     }
 
     @Test
     public void test_name() {
-        assertEquals(HierarchicalPageProperty.SOURCE, injector.getName());
+        assertEquals(HierarchicalPagePropertyInjector.SOURCE, injector.getName());
     }
 
     @Test
     public void test() {
-        assertNotNull(adapted);
+        assertNotNull(hierarchicalModel);
+        assertEquals("inherited!", hierarchicalModel.getHierarchicalPagePropertyString());
+        assertNull(pageModel.getHierarchicalPagePropertyString());
+        assertEquals("not inherited", hierarchicalModel.getPagePropertyString());
     }